Added the functionality to post data via vue js and save in the database

+++ database/migrations/2017_10_04_071643_create_phonbooks_table.php
@@ -16,7 +16,7 @@ class CreatePhonbooksTable extends Migration
         Schema::create('phonbooks', function (Blueprint $table) {
             $table->increments('id');
             $table->string('name');
-            $table->integer('phoneno');
+            $table->bigInteger('phoneno');
             $table->string('email');
             $table->timestamps();
         });
